Use when/2 instead of freeze/2
When using the/2 with DCGs, it's quite common for a value to be a list
whose tail is a variable.  That's fine but prevented type assertions
like the(codes,X).  As soon as X was bound to a partial list, we'd run
the type check and it failed.

Instead of triggering type checks on unification, wait until the value
is fully ground.  In nearly every case, error:has_type/2 definitions
require the arguments to be ground.  This change breaks the few
remaining cases.  If it becomes a problem, I can fix it once I have
some concrete examples.