Avoid singleton warnings
Recent versions of SWI-Prolog have "semantic singleton" detection.  It
noticed that Slash was a singleton with respect to each possible flow of
control. I don't want to remove Slash because it seems likely that we'll
use it sometime in the future.  It also serves to remind me of the
structure of the mode term.